Add options to disable automatic scheduling and failover of BGP speakers
Sometimes it is not desirable to reschedule BGP speakers to
another DrAgents, especially in cases when the speaker must
run on a specific DrAgent in a specific place in the network
topology. This commit adds configuration flags which allow to
disable automatic scheduling of new BGP speakers on DrAgents,
and  rescheduling in case of DrAgent failure, so the speaker
is bound to the same DrAgent, even if the the agent is down.

Change-Id: I5364749e7a4009230a8b8174ad599cf40a24f415